Move next and auth monitor elements to tools dir.
